Update Spinely.ino
When arduino code receives command from php, determine which function (calibration, monitoring, turn_on_motor). When arduino code receives function it is "operating", otherwise it is "waiting".

this Arduino device starts off in this chill state where it's just waiting around, not really doing anything important (aka the "isOperating = false" state). Every now and then, the device wakes up and goes, "Hey, is there anything new from that PHP server?" It does this using "checkForCommands()".

If the device actually gets a command from the server, it sets its state to "isOperating = true" (meaning it's ready to roll). Based on the command it received, the device picks the right function to execute.